requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openpyxl
pyjwt
passlib
bcrypt==3.2.0
python-multipart
psycopg2-binary